Refactor batch and state into blueprints
