+++ app/utils.py
@@ -1,0 +1,12 @@
+from datetime import datetime
+
+import pytz
+from flask import jsonify
+
+
+def api_error(message, status_code=400):
+    return jsonify({"error": message}), status_code
+
+
+def updated_at():
+    return datetime.now(pytz.utc).replace(microsecond=0).isoformat()
